+++ src/ask_for_help_bot/state_mixin.py
@@ -21,6 +21,7 @@ class StateMixin:
     STATE_ANSWERING_SENSITIVE = "answer_sensitive"
     STATE_ANSWERING_ANONYMOUSLY = "answer_anonymously"
     STATE_BEST_ANSWER_0 = "best_answer_0"
+    STATE_BEST_ANSWER_PUBLISH = "best_answer_publish"
     STATE_BEST_ANSWER_1 = "best_answer_1"
 
     def _is_doing_another_action(self, context: ConversationContext) -> bool:
@@ -31,7 +32,7 @@ class StateMixin:
             self.STATE_ANSWERING, self.STATE_ANSWERING_SENSITIVE, self.STATE_ANSWERING_ANONYMOUSLY,
             self.STATE_QUESTION_0, self.STATE_QUESTION_1, self.STATE_QUESTION_2, self.STATE_QUESTION_3,
             self.STATE_QUESTION_4, self.STATE_QUESTION_4_1, self.STATE_QUESTION_5, self.STATE_QUESTION_6,
-            self.STATE_BEST_ANSWER_0, self.STATE_BEST_ANSWER_1
+            self.STATE_BEST_ANSWER_0, self.STATE_BEST_ANSWER_PUBLISH, self.STATE_BEST_ANSWER_1
         ]
         current_status = context.get_static_state(self.CONTEXT_CURRENT_STATE, "")
         return current_status in statuses